Rename HttpLogAlertBuilder
